requirements.txt
fastapi
uvicorn
python-dotenv
pyjwt
httpx
pydantic
pyjwt
starlette
pydantic-settings
